atkplug: Add atk_plug_new_child
As discussed on

https://mail.gnome.org/archives/gnome-accessibility-devel/2019-August/msg00001.html

in some cases such as gtk, one can't use the AtkPlug type directly as
accessible object for the toplevel widget of the application, and one
has to make AtkPlug a parent of the standard accessible object for that
toplevel widget.

atk_plug_new_child() allows to achieve this easily by making the new
AtkPlug a parent of the child passed as parameter.