build: Improve bind_textdomain_codeset() check
The bind_textdomain_codeset() function may not be present in the CRT,
but in a separate libintl (gettext-runtime) library, or loaded via
proxy-intl (noted as "libintl" below).

Based on the way it is done in GLib: this looks for the ngettext()
function from the CRT, and if it is not found, we want to look into
libintl to check for the presence of this function.

Note that since GLib links to libintl if gettext-runtime functions are
not found in the CRT and does so in its pkg-config file, we do not
explicitly link to it here when building libatk.